==> step1_Complete Trajectory Adjacency list preparation for Project Type.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os


path =  # Enter the directory of your project folder
PATH = path + "/Results"

if not os.path.exists(PATH):
    os.mkdir(PATH) 




# Dataframe preparation
df = pd.read_csv(path+"//ToyDataset.csv").drop(['Unnamed: 0'],axis=1)

clientID = df["ClientID"].unique()

edges = pd.DataFrame([],columns = ["start","end"])
for i in range(len(clientID)):
    temp = df[df["ClientID"]==clientID[i]].reset_index(drop=True)
    for k in range(len(temp)-1):
        start = temp['ProjectType'].iloc[k]
        end = temp['ProjectType'].iloc[k+1]
        edges = edges.append({"start":start,"end":end,"edge": (start,end)},ignore_index=True)
    


adj = edges.groupby(['start','end']).count()

p = adj.index.get_level_values(0).unique().values
adj_list = pd.DataFrame([],columns = ['source',"target","weight"])

for i in p:
    l = []
    for k in range(len(adj)):
        if adj.iloc[k].name[0] == i:
            adj_list = adj_list.append({"source":i,"target":adj.iloc[k].name[1],"weight":adj.iloc[k].values[0]},ignore_index=True)
    
    

edge_list = pd.DataFrame()
for i in p:
    temp = adj_list[adj_list["source"]==i] 
    total = temp['weight'].sum()
    temp['weight'] = temp['weight']/total
    edge_list = edge_list.append(temp,ignore_index=True)
    
edge_list.to_csv(PATH+"//Adjacency List ProjectType.csv")
