==> step2_One hot encoding.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os


path =  #  # Enter the directory of your project folder
PATH = path + "/Results"

if not os.path.exists(PATH):
    os.mkdir(PATH) 




# Dataframe preparation
df = pd.read_csv(path+"//ToyDataset.csv").drop(['Unnamed: 0'],axis=1)

clientID = df["ClientID"].unique()

# binning the benefit total, member number, and children number

df['Benefit total']
bins = [-0.5, 2000,8000]
labels = [1,2]

df['Benefit total_binned'] = pd.cut(df['Benefit total'],  bins=bins, labels=labels)

df['memberNo']
bins = [0,2,7]
labels = ["small","big"]

df['memberNo_binned'] = pd.cut(df['memberNo'],  bins=bins, labels=labels)

df['childrenNo']
bins = [-1,1,7]
labels = ["small","big"]

df['childrenNo_binned'] = pd.cut(df['childrenNo'],  bins=bins, labels=labels)

df = df.drop(['Benefit total','memberNo','childrenNo'],axis=1)

data = df.drop(['EnrollmentID', 'ClientID', 'ProjectID', 'EntryDate', 'ProjectType',
       'ExitID', 'ExitDate','HouseholdID'],axis = 1)
cols = data.columns
data_final = pd.DataFrame([])
for i in cols:
    x = pd.get_dummies(data[i],prefix=i)
    data_final = pd.concat([data_final,x],axis=1)
data = df.drop(cols,axis=1)
data = pd.concat([data,data_final],axis=1)
data.to_csv(PATH+"//Data one hot encoded.csv")
